Add files via upload
Differences between Version 1 (centos_fresh_v1.sh) and Version 2 (this version; centos_fresh_v2.sh):

COMMAND OUTPUT REDIRECTION AGGREGATION: In Version 1 of this script (centos_fresh_v1.sh), multiple commands were executed and at the end of each execution, output redirection was performed.  Change was made to group all relative commands into braces and, in turn, take those outputs and redirect one time per grouping.  So, rather than redirecting after each command, all commands for a particular section are executed and their outputs are redirected once (after they have all been executed).

DOCKER INSTALL: In the "Docker Removal and Install" section, during the Docker install, an "if/else" statement is ran to install Docker.  Change was made to better define the variable in the "if/else" statement.  Additonally, the Docker service was set to be enabled in this portion of the script, so that upon reboot, the service will be running.  NOTE: Docker will show as "inactive" during the service verifications prior to reboot.

DOCKER REMOVAL: In Version 1 of this script (centos_fresh_v1.sh), the Docker removal section spelled out, line-by-line, each package.  This version (Version 2) uses the wildcard (*) for Docker in the "Docker Removal and Install" section to uninstall Docker packages, which consolidates the code by several lines.

ECHO: In Version 1 of this script (centos_fresh_v1.sh), "echo" was used abundantly.  While it made for a more readable and cosmetically appealing log file, it took up a lot of space in the script.  "echo" has been significantly reduced in this script.

GLOBS: In the "LibreOffice Removal and Install" section, there is a line of code to install the RPM packages that are extracted from the downloaded LibreOffice file.  Change was made to make sure that files containing dashes (-) are not interpreted as command flags/options.  Additionally, variables used in the "Update System and Verify Services" section were changed to have double quotes to prevent globbing.

LIBREOFFICE: In the "LibreOffice Removal and Install" section, there is a line of code in which the directory needs to be changed into the newly downloaded and extracted LibreOffice directory.  A change was made to provide protection in case there is a failure in executing "cd" to that appropriate directory.

SPINNER FUNCTION: In Version 1 of this script (centos_fresh_v1.sh), the spinner function was called everytime a removal or install was initiated and separate "kill" commands were executed at the end of those removals or installs.  This version (Version 2) starts the spinner at the beginning of the script (starts in the "Docker Removal and Install" section) and then stores the spinner process in a variable.  THEN, at the end, the spinner variable (spinner process, really) gets killed.  This removed several lines of unneccessary code.

USER ADDITION: In the "Docker Removal and Install" section, there is a portion of code during the install that configures a Docker user.  The default user in the script is "dockman".  Comments were added to elaborate on this user.  If the user is not desired/needed, this portion of the code/script can be either removed or commented out.  If a different username is preferred, then appropriate changes can be made.  This is entirely user discretion.

VERBIAGE: Changed verbiage in both comments and statements echoed to users on terminal to be more "standard" and simple.  The description of this script was also modified.

VARIABLES: Removed the "$bold" and $normal" variables.  They ultimately were not needed and did not appear to function consistently/persistently.